requirements.txt
streamlit
langchain-openai
langchain
ipykernel
python-dotenv
langchain_community
langchain_core
openai
langchain-groq
ollama
pymupdf
langchain.text-splitters
langchain_huggingface
faiss-cpu
langchain_chroma
arxiv
wikipedia
duckduckgo-search


